Get the map scrolling code working when moving left
Basically the idea is that the background isn't just limited to the viewport;
there's actually a buffer of tiles (currently 10 on each side) around the part
of the map that the viewport frames. When the player crosses over one of the
bounds of the background, we load the part of the map the player is now in which
is not in the buffer, and we add it to the buffer. The buffer always stays the
same size.

Right now this is only working if you move left.