Trying to make the game loop / draw code faster
One thing we are doing is to ensure that every tick happens at a regular
interval. The thing about the game loop is that we don't know how long each tick
will take to complete. We hope that each tick will take an insignificant amount
of time, but that may not happen.
So for instance, one iteration may take 1ms, the next may take 3ms, the next
2ms, the next 5ms, etc. That is all fine as long as we can control when each
tick *starts*.

There are two ways to run the game loop: requestAnimFrame and setTimeout. Using
requestAnimFrame supposedly saves CPUs cycles, but it doesn't seem to work as
well as you might think. Since it fires as fast as possible, you end up having
to limit the FPS. And since repeating the game loop happens automatically, you
can't control when each tick starts.

Using setTimeout prevents you from having to control the FPS (since by
specifying an interval you cap the FPS). This solution isn't perfect, either,
though. An interval of 30ms doesn't mean each tick is guaranteed to *take* 30ms,
only that the next tick will be scheduled for 30ms after this tick ends; how
long each tick takes is still variable. (We can think of the game loop like tick
+ wait + tick + wait + …) To compensate for this, we can adjust the wait time
after a tick runs based on its duration to ensure that the total time of tick +
pause is a consistent amount.