Initial code (and restructuring) to add an enemy
We would like to restrict the position of an enemy to a box, but we already do
this for the player with the viewport. So the 'fence' is now in Mob instead of
Player.

There's some new collision code to keep an enemy inside the fence and outside
collision boxes (the code to determine this is simpler with enemies than with
the player).

Bounds is now defined as (width, height, x1, y1) instead of (x1, y1, x2, y2) so
that proportionally scaling and moving by corner is straightforward. (We always
create new Bounds with a pre-existing width and height, anyway.)

Augment the sprite stuff to support animation sequences which are transitions --
for instance, when an enemy is turning right to down, the "right-to-down"
animation plays for 2 frames, then the "down" animation plays.