LOTS of refactorings
We now use a custom fabricator to define classes and modules based on John
Resig's Simple Inheritance pattern instead of CoffeeScript's built-in class
support. This was necessary so that reusing code using modules was easier. To
wit: since we tend to use bundles of the same methods all over the place, we've
collected these into "traits" -- interfaces defined as modules that we can mix
in to classes or other modules.

The core draw loop and behavior have been extracted into a separate file,
naturally called "core". Core, the keyboard module, the viewport, and other
modules now know how to plug directly into the main module at different points
-- this way it is easy to disable a single plugin (say, the FPS reporter).

Some of the methods in Mob have been split off into a simpler class called
"Grob". The idea is that all grobs are drawable and perhaps collidable, but mobs
have specific behavior that applies to moving objects, such as the player,
NPC's, or even animated sprites.

Previously, CollisionBoxes was a collection of CollisionBox which wrapped either
a Bounds (used to delineate collidable areas of the map) or Mob (like the player
or an enemy). Now, CollidableCollection is a collection of Grob, which may be
one of Player, Enemy or MapBlock, and has a @box attribute that points to a
CollidableBox which wraps the Grob's bounds.