Yay, we now have intra-layer object layering
If the player is above an object on the map (above meaning y is less), then it
is drawn before the object is drawn. As soon as the player moves below the
object, then the object is drawn before the player. This works because objects
are drawn from top to bottom, left to right, and are kept in the order in which
they are to be drawn. Every time the player moves, we re-evaluate where it
should appear in this list based on its position on the map.